requirements.txt
pydantic~=1.8.2
PyJWT~=2.3.0
fastapi~=0.70.1
pymongo~=4.1.1
starlette~=0.16.0
emails~=0.6
requests~=2.26.0
itsdangerous~=2.0.1
APScheduler~=3.8.1
python-dotenv~=0.19.2
uvicorn~=0.16.0
loguru~=0.5.3
dnspython~=2.2.1
loggly-python-handler~=1.0.1
email-validator~=1.1.3
jinja2~=3.0.3
google~=3.0.0
sentry-sdk~=1.5.1
google-auth~=2.3.3
pytest~=6.2.5
nox~=2021.10.1
poetry~=1.1.12
isort~=5.10.1
black~=21.12b0
web3~=5.25.0
py-solc-x~=1.1.1
frozendict
mnemonic